started to set up react router routes. Created server routes and routes file for cleaner files
